Change transfer output to long from decimal

+++ dngn-api-backend/Integrations/Models/Common/BaseTransferOutput.cs
@@ -7,8 +7,8 @@ namespace DngnApiBackend.Integrations.Models.Common
         public string? StatusMessage { get; set; }
         public TransactionStatus Status { get; set; }
         public string? AccountNumber { get; set; }
-        public decimal Amount { get; set; }
-        public decimal TransferFees { get; set; }
+        public long Amount { get; set; }
+        public long TransferFees { get; set; }
         public string? ProviderTransactionReference { get; set; }
         public string? ProviderBankReference { get; set; }
     }
